Reimplemented the counting of calls to oracles in the crypto transformation + minor changes
Merge of branch 2021.16.fix_proba_bug

Reimplemented the counting of calls to oracles in the crypto
transformation, to make it more precise and fix two bugs.
- when computing a replication bound, exploit that we count the number
of calls or random generations *for each random variable chosen above
the replication*, hence these random variables must be the same: same
names, same indices.
- use defined variables (in addition to known facts) to show that
two oracle calls are incompatible. For that, compute the incompatibility
information between program points; that leads to removing some
useless branches in the output the crypto transformation.
- in incompatibility tests between oracles, make sure that indices
are mapped in a coherent way.
- when trying to eliminate indices, prove that the oracle call is the
same (instead of proving that all indices are the same) assuming fixed
counted indices and possibly different uncounted indices. That allows
to eliminate more counted indices.
- also apply these elimination of indices and incompatibility tests to
the counting of creations of random variables.
- count the creations of random variables, do not try to count how
many are used in oracles, so count them at their creation point.
(That's cleaner.)
- Improvements in detection of subsumed oracle calls
  * Use known facts at oracle2 to perform the matching.
  * Improved matching modulo both the equational theory and known equalities
    from the protocol.
  * Do not redo the filtering of indices when it removed nothing.
  * When several oracle calls are subsumed by the newly added oracle call,
    remove all of them.
- do not consider as incompatible oracle calls that have probabilities
#O for different O, even when they are in different branches: these
probabilities must be added.
(Bug exhibited in examplesnd/otest/proba_countO_diffbranch2.ocv)
- fixed bug: when an oracle call O1 was detected as being an instance
of O2, the compability information for O2 was correctly kept, but
sometimes the probability for O1 was kept instead of the probability
for O2.
(Bug exhibited in examplesnd/otest/bug_same_oracle_call.ocv)
- optimized the algorithm by using an algorithm for computing the
maximum independent set in a graph.

- In Diffie-Hellman macros, when there is an oracle ODHa9(x) = exp(g,
mult(a,x)), add #ODHa9 * time(exp) to the runtime of the context. Same
for ODHb9.
- In GDH macros, avoid counting m = exp(g,a) as a call to
m = exp(m',a) (a call to the DDH oracle), since it can be computed by
comparing with the public key. Done by adding a specific oracle for
m = exp(g,a)
- HPKE: allow eliminating a bit more collisions, to get nicer
probabilities (DDH queries in O(qH) in the application of GDH)
- Improved probabilities in OEKE, by a more precise manual guidance.