Allow proving that a find/get is [unique] after some game transformations
Merge of branch 2022.15.prove_unique_later

Removed setting allowUnprovedUnique.
This extension required some subtle changes:
- computation of probabilities:
    - If we do not prove secrecy, Adv_Q(C,||_i D_i) <= sum_i Adv_Q(C,D_i)
    - If we prove secrecy, Adv_Q(C, ||_i D_i) <=
        sum_i Adv_Q(C,D_i) + sum_{i \neq i_0} \Adv_Q(C, D_{NUi})
      where D_{i_0} is the element that contains the secrecy query and
      D{NUi} contains only the queries for non-unique events in D_i.
      The additional term sum_{i \neq i_0} \Adv_Q(C, D_{NUi}) may be surprising.
- a find[unique] not proved yet may abort (fct Terms.may_abort_counted).
- consider a stronger version of may_abort in which find[unique] may
abort even with it is proved, although these traces are not counted in
the probability (fct Terms.may_abort).
- disable some game transformations for find[unique] when unicity
is not proved yet, although these transformations are allowed for
non-unique find (merging of branches, simplification of single branch
with not index).