Fixed 2 bugs in the pre-treatment of equiv statements.
Merge of branch 2023.05.bugfix

- Fixed internal error when the current replication indices in the RHS
of an equiv statement are shorter than the suffix of the maximum
sequence of indices of variables in the defined condition obtained by
removing the indices of find.
- Correctly check that variables in the RHS that use indices provided
as arguments to the oracle match the corresponding variables in the LHS,
to make sure that they will be defined after the crypto transformation.